Remove Leaflet code
